[add] HeaderNavi component
